Vascular Density, Dark patches.
The above code can be used to find out the localised vascular density of an layer 1 or layer 2 image obtained from OCT scans. It can also detect the number of dark patches present in the image. The above code can be used for healthy and diseased eyes, the following is found out using a statistical approach.